[DTensor] Refactor strategy/rule registration into dedicated module
To avoid circular import issues:
- utils.py used to include registration functions which import/depend on
  DTensor.sharding_propagator
- I plan to use other utils from utils.py inside sharding_propagator.py

[ghstack-poisoned]
