+++ torch/distributed/tensor/_ops/_random_ops.py
@@ -6,7 +6,8 @@ from torch.distributed.tensor._op_schema import (
     OpStrategy,
     StrategyType,
 )
-from torch.distributed.tensor._ops.utils import is_tensor_partial, register_op_strategy
+from torch.distributed.tensor._ops.registration import register_op_strategy
+from torch.distributed.tensor._ops.utils import is_tensor_partial
 
 
 aten = torch.ops.aten
